refactor : changing id column to 'Long' and columns with 'Date' to 'LocalDateTime'

+++ src/main/java/com/system_management_student/system_management_student/modal/dto/ViewData/DataUsersDto.java
@@ -7,6 +7,7 @@ import lombok.Builder;
 import lombok.Data;
 
 import java.io.Serializable;
+import java.time.LocalDateTime;
 import java.util.Date;
 import java.util.List;
 import java.util.stream.Collectors;
@@ -19,7 +20,7 @@ import java.util.stream.Collectors;
 @JsonInclude(JsonInclude.Include.NON_NULL)
 @JsonIgnoreProperties(ignoreUnknown = true)
 public class DataUsersDto implements Serializable {
-    private Integer id;
+    private Long id;
     private String name;
     private String email;
     private String rg;
