+++ src/main/java/com/system_management_student/system_management_student/modal/entity/DataUsers.java
@@ -3,6 +3,7 @@ package com.system_management_student.system_management_student.modal.entity;
 import jakarta.persistence.*;
 import lombok.Data;
 
+import java.time.LocalDateTime;
 import java.util.Date;
 import java.util.List;
 
@@ -11,7 +12,7 @@ import java.util.List;
 public class DataUsers {
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-    private Integer id;
+    private Long id;
 
     @Column(name = "name")
     private String name;
